+++ backend/src/main/java/com/nbe3/cafemanagement/controller/OrderController.java
@@ -1,0 +1,4 @@
+package com.nbe3.cafemanagement.controller;
+
+public class OrderController {
+}
