+++ backend/src/main/java/com/nbe3/cafemanagement/controller/OrderDetailController.java
@@ -1,0 +1,4 @@
+package com.nbe3.cafemanagement.controller;
+
+public class OrderDetailController {
+}
