+++ backend/src/main/java/com/nbe3/cafemanagement/domain/Admin.java
@@ -1,0 +1,25 @@
+package com.nbe3.cafemanagement.domain;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.LocalDateTime;
+
+@Getter
+@Setter
+@Entity
+public class Admin {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private int id;
+
+    private LocalDateTime created_at;
+
+    private String username;
+
+    private String password;
+}
