+++ backend/src/main/java/com/nbe3/cafemanagement/domain/OrderDetail.java
@@ -1,14 +1,8 @@
-package domain;
+package com.nbe3.cafemanagement.domain;
 
 import jakarta.persistence.*;
 import lombok.*;
-import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.Date;
-import java.util.List;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
