+++ backend/src/main/java/com/nbe3/cafemanagement/dto/MainRequestDto.java
@@ -1,0 +1,4 @@
+package com.nbe3.cafemanagement.dto;
+
+public class MainRequestDto {
+}
