+++ backend/src/main/java/com/nbe3/cafemanagement/repository/OrderDetailRepository.java
@@ -1,4 +1,4 @@
-package repository;
+package com.nbe3.cafemanagement.repository;
 
 public interface OrderDetailRepository {
 }
