+++ backend/src/main/java/com/nbe3/cafemanagement/repository/OrderRepository.java
@@ -1,0 +1,4 @@
+package com.nbe3.cafemanagement.repository;
+
+public interface OrderRepository {
+}
