+++ backend/src/main/java/com/nbe3/cafemanagement/service/OrderDetailService.java
@@ -1,0 +1,4 @@
+package com.nbe3.cafemanagement.service;
+
+public class OrderDetailService {
+}
