+++ backend/src/main/java/com/nbe3/cafemanagement/service/OrderService.java
@@ -1,0 +1,4 @@
+package com.nbe3.cafemanagement.service;
+
+public class OrderService {
+}
