--- backend/src/main/java/controller/MainController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class MainController {
-}
