--- backend/src/main/java/controller/OrderController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class OrderController {
-}
