--- backend/src/main/java/controller/OrderDetailController.java
@@ -1,4 +1,0 @@
-package controller;
-
-public class OrderDetailController {
-}
