--- backend/src/main/java/domain/Admin.java
@@ -1,11 +1,0 @@
-package domain;
-
-import jakarta.persistence.Entity;
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-@Entity
-public class Admin {
-}
