--- backend/src/main/java/dto/MainRequestDto.java
@@ -1,4 +1,0 @@
-package dto;
-
-public class MainRequestDto {
-}
