--- backend/src/main/java/dto/OrderDetailDto.java
@@ -1,4 +1,0 @@
-package dto;
-
-public class OrderDetailDto {
-}
