--- backend/src/main/java/dto/OrderDto.java
@@ -1,4 +1,0 @@
-package dto;
-
-public class OrderDto {
-}
