--- backend/src/main/java/repository/MainRepository.java
@@ -1,4 +1,0 @@
-package repository;
-
-public class MainRepository {
-}
