--- backend/src/main/java/repository/OrderRepository.java
@@ -1,4 +1,0 @@
-package repository;
-
-public interface OrderRepository {
-}
