--- backend/src/main/java/service/MainService.java
@@ -1,4 +1,0 @@
-package service;
-
-public class MainService {
-}
