--- backend/src/main/java/service/OrderDetailService.java
@@ -1,4 +1,0 @@
-package service;
-
-public class OrderDetailService {
-}
