--- backend/src/main/java/service/OrderService.java
@@ -1,4 +1,0 @@
-package service;
-
-public class OrderService {
-}
