+++ backend/src/main/resources/application.properties
@@ -1,28 +1,28 @@
 spring.application.name=cafemanagement
 
-# Thymeleaf 설정
+# Thymeleaf ??
 spring.thymeleaf.prefix=classpath:/templates/
 spring.thymeleaf.suffix=.html
 spring.thymeleaf.mode=HTML5
 spring.thymeleaf.cache=false
 
-# 서버 포트 (원하는 포트로 설정 가능)
+# ?? ?? (??? ??? ?? ??)
 server.port=9090
 
-# 로컬 MySQL 데이터베이스 연결 설정
-spring.datasource.url=jdbc:mysql://localhost:3306/coffee09?useSSL=false&serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=lldj123414
+# ?? MySQL ?????? ?? ??
+spring.datasource.url=jdbc:mysql://DESKTOP-BGJMA8A:13307/nbe_3_4_team09?useSSL=false&serverTimezone=UTC
+spring.datasource.username=nrkim
+spring.datasource.password=1234
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 
-# 원격 MySQL 데이터베이스 연결 설정
+# ?? MySQL ?????? ?? ??
 # spring.datasource.url=jdbc:mysql://sql12.freesqldatabase.com:3306/sql12757316?useSSL=false&serverTimezone=UTC
 # spring.datasource.username=sql12757316
 # spring.datasource.password=qSBrXV4vgL
 # spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
-# JPA 설정
+# JPA ??
 spring.jpa.hibernate.ddl-auto=update
 spring.jpa.show-sql=true
 spring.jpa.properties.hibernate.format_sql=true
