--- backend/src/main/resources/static/css/init.css
@@ -1,3 +1,0 @@
-#test{
-    background-color: crimson;
-}
