+++ backend/src/main/resources/templates/main.html
@@ -8,66 +8,10 @@
   <!-- Bootstrap CSS -->
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/css/bootstrap.min.css" rel="stylesheet"
         integrity="sha384-KyZXEAg3QhqLMpG8r+8fhAXLRk2vvoC2f3B09zVXn8CA5QIVfZOJ3BCsw2P0p/We" crossorigin="anonymous">
-  <style>
-      body {
-          background: #ddd;
-      }
 
-      .card {
-          margin: auto;
-          max-width: 950px;
-          width: 90%;
-          box-shadow: 0 6px 20px 0 rgba(0, 0, 0, 0.19);
-          border-radius: 1rem;
-          border: transparent
-      }
+  <!-- Custom CSS -->
+  <link href="styles.css" rel="stylesheet">
 
-      .summary {
-          background-color: #ddd;
-          border-top-right-radius: 1rem;
-          border-bottom-right-radius: 1rem;
-          padding: 4vh;
-          color: rgb(65, 65, 65)
-      }
-
-      @media (max-width: 767px) {
-          .summary {
-              border-top-right-radius: unset;
-              border-bottom-left-radius: 1rem
-          }
-      }
-
-      .row {
-          margin: 0
-      }
-
-      .title b {
-          font-size: 1.5rem
-      }
-
-      .col-2,
-      .col {
-          padding: 0 1vh
-      }
-
-      img {
-          width: 3.5rem
-      }
-
-      hr {
-          margin-top: 1.25rem
-      }
-      .products {
-          width: 100%;
-      }
-      .products .price, .products .action {
-          line-height: 38px;
-      }
-      .products .action {
-          line-height: 38px;
-      }
-
-  </style>
   <title>Hello, world!</title>
 </head>
 <body class="container-fluid">
@@ -114,7 +58,7 @@
       </div>
       <hr>
       <div class="row">
-        <h6 class="p-0">Columbia Nariñó <span class="badge bg-dark text-">2개</span></h6>
+        <h6 class="p-0">Columbia Nariñó <span class="badge bg-dark">2개</span></h6>
       </div>
       <div class="row">
         <h6 class="p-0">Brazil Serra Do Caparaó <span class="badge bg-dark">2개</span></h6>
